refactor: move dependency injection to a separate project
